backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==4.0.1
pydantic-settings
email-validator
python-multipart
google-auth
requests
